Clean up and organize ProposalsTable component
- Moved logic to the table component
- Added clarification comments
- Moved skeleton to own file
- Simplified the sort handling logic
