+++ apps/council-ui/src/ui/proposals/ProposalsTable.tsx
@@ -1,8 +1,7 @@
 import { Ballot } from "@council/sdk";
 import Link from "next/link";
 import { useRouter } from "next/router";
-import { ReactElement } from "react";
-import Skeleton from "react-loading-skeleton";
+import { ReactElement, useMemo, useState } from "react";
 import { makeProposalURL } from "src/routes";
 import { ChevronRightSVG } from "src/ui/base/svg/ChevronRight";
 import { DownArrowSVG } from "src/ui/base/svg/DownArrow";
@@ -10,27 +9,37 @@ import { UpArrowSVG } from "src/ui/base/svg/UpArrow";
 import { MiniQuorumBar } from "src/ui/proposals/MiniQuorumBar";
 
 export interface ProposalRowData {
-  votingContractName: string;
-  votingContractAddress: string;
-  id: number;
-  created: Date | null;
-  votingEnds: Date | null;
-  currentQuorum: string;
-  requiredQuorum: string | null;
   ballot: Ballot | null;
+  created: Date | null;
+  currentQuorum: string;
+  id: number;
+  requiredQuorum: string | null;
+  votingContractAddress: string;
+  votingContractName: string;
+  votingEnds: Date | null;
 }
 
 interface ProposalsTableProps {
-  rowData?: ProposalRowData[];
-  sortOptions: ProposalSortOptions;
-  onSortOptionsChange: (field: ProposalSortField) => void;
+  rowData: ProposalRowData[];
 }
 
-export function ProposalsTable({
-  rowData,
-  sortOptions,
-  onSortOptionsChange,
-}: ProposalsTableProps): ReactElement {
+export function ProposalsTable({ rowData }: ProposalsTableProps): ReactElement {
+  const [sortOptions, setSortOptions] = useState<ProposalSortOptions>({
+    field: ProposalSortField.ENDS,
+    direction: ProposalSortDirection.ASC,
+  });
+
+  const sortedData = useMemo(
+    () => sortProposalRowData(sortOptions, rowData),
+    [sortOptions, rowData],
+  );
+
+  const handleSortOptionsChange = (field: ProposalSortField) =>
+    setSortOptions({
+      field,
+      direction: sortStates[sortOptions.direction],
+    });
+
   return (
     <table className="w-full daisy-table-zebra daisy-table min-w-fit">
       <thead>
@@ -39,161 +48,68 @@ export function ProposalsTable({
 
           <th
             className="cursor-pointer"
-            onClick={() => onSortOptionsChange(ProposalSortField.ENDS)}
+            onClick={() => handleSortOptionsChange(ProposalSortField.ENDS)}
           >
             <span className="mr-1 select-none hover:underline">
               Voting Ends
             </span>
-            {sortOptions.field === ProposalSortField.ENDS && (
-              <SortDirectionStatus direction={sortOptions.direction} />
-            )}
+            <SortDirectionStatus
+              direction={sortOptions.direction}
+              enabled={sortOptions.field === ProposalSortField.ENDS}
+            />
           </th>
 
           <th
             className="cursor-pointer"
-            onClick={() => onSortOptionsChange(ProposalSortField.QUORUM)}
+            onClick={() => handleSortOptionsChange(ProposalSortField.QUORUM)}
           >
             <span className="mr-1 select-none hover:underline">Quorum</span>
-            {sortOptions.field === ProposalSortField.QUORUM && (
-              <SortDirectionStatus direction={sortOptions.direction} />
-            )}
+            <SortDirectionStatus
+              direction={sortOptions.direction}
+              enabled={sortOptions.field === ProposalSortField.QUORUM}
+            />
           </th>
 
           <th>Your Ballot</th>
 
+          {/* empty header reserved for button */}
           <th></th>
         </tr>
       </thead>
 
       <tbody>
-        {rowData &&
-          rowData.map((proposal) => (
-            <ProposalTableRow
-              {...proposal}
-              key={`${proposal.votingContractName}${proposal.id}`}
-            />
-          ))}
-      </tbody>
-    </table>
-  );
-}
-
-export function ProposalTableSkeleton(): ReactElement {
-  return (
-    <table className="w-full shadow-md daisy-table-zebra daisy-table min-w-fit">
-      <thead>
-        <tr>
-          <th className="w-72">Name</th>
-
-          <th className="w-32">Voting Ends</th>
-
-          <th className="w-32">
-            <span className="mr-1">Quorum</span>
-          </th>
-
-          <th className="w-16">Your Ballot</th>
-
-          <th className="w-16"></th>
-        </tr>
-      </thead>
-
-      <tbody>
-        <tr>
-          <th>
-            <Skeleton />
-          </th>
-          <td>
-            <Skeleton />
-          </td>
-          <td>
-            <Skeleton />
-          </td>
-          <td>
-            <Skeleton />
-          </td>
-          <td>
-            <Skeleton />
-          </td>
-        </tr>
-
-        <tr>
-          <th>
-            <Skeleton />
-          </th>
-          <td>
-            <Skeleton />
-          </td>
-          <td>
-            <Skeleton />
-          </td>
-          <td>
-            <Skeleton />
-          </td>
-          <td>
-            <Skeleton />
-          </td>
-        </tr>
-
-        <tr>
-          <th>
-            <Skeleton />
-          </th>
-          <td>
-            <Skeleton />
-          </td>
-          <td>
-            <Skeleton />
-          </td>
-          <td>
-            <Skeleton />
-          </td>
-          <td>
-            <Skeleton />
-          </td>
-        </tr>
-
-        <tr>
-          <th>
-            <Skeleton />
-          </th>
-          <td>
-            <Skeleton />
-          </td>
-          <td>
-            <Skeleton />
-          </td>
-          <td>
-            <Skeleton />
-          </td>
-          <td>
-            <Skeleton />
-          </td>
-        </tr>
+        {sortedData.map((proposal) => (
+          <ProposalTableRow
+            {...proposal}
+            key={`${proposal.votingContractName}${proposal.id}`}
+          />
+        ))}
       </tbody>
     </table>
   );
 }
 
 function ProposalTableRow({
+  ballot,
+  currentQuorum,
+  id,
+  requiredQuorum,
   votingContractAddress,
   votingContractName,
-  id,
   votingEnds,
-  currentQuorum,
-  requiredQuorum,
-  ballot,
 }: ProposalRowData) {
   const { push } = useRouter();
-
   return (
     <tr
       onClick={() => push(makeProposalURL(votingContractAddress, id))}
       className="hover:cursor-pointer"
     >
-      <th>
+      <td>
         {votingContractName} Proposal {id}
-      </th>
+      </td>
+
       <td>{votingEnds?.toLocaleDateString() ?? <em>unknown</em>}</td>
+
       <td>
         <MiniQuorumBar
           currentQuorum={currentQuorum}
@@ -201,16 +117,38 @@ function ProposalTableRow({
           votingEnds={votingEnds}
         />
       </td>
+
       <td>{ballot ?? <em>Not voted</em>}</td>
-      <th>
+
+      <td>
         <button className="daisy-btn daisy-btn-ghost daisy-btn-sm">
           <Link href={makeProposalURL(votingContractAddress, id)}>
             <ChevronRightSVG />
           </Link>
         </button>
-      </th>
+      </td>
     </tr>
   );
+}
+
+interface SortDirectionProps {
+  direction: ProposalSortDirection;
+  enabled: boolean;
+}
+
+function SortDirectionStatus({ direction, enabled }: SortDirectionProps) {
+  if (!enabled) {
+    return null;
+  }
+
+  switch (direction) {
+    case ProposalSortDirection.ASC:
+      return <DownArrowSVG />;
+    case ProposalSortDirection.DESC:
+      return <UpArrowSVG />;
+    default:
+      return null;
+  }
 }
 
 export enum ProposalSortField {
@@ -229,17 +167,38 @@ export interface ProposalSortOptions {
   direction: ProposalSortDirection;
 }
 
-function SortDirectionStatus({
-  direction,
-}: {
-  direction: ProposalSortDirection;
-}) {
-  switch (direction) {
-    case ProposalSortDirection.ASC:
-      return <DownArrowSVG />;
-    case ProposalSortDirection.DESC:
-      return <UpArrowSVG />;
-    default:
-      return null;
+// simple state machine for sort state transitions
+const sortStates: Record<ProposalSortDirection, ProposalSortDirection> = {
+  [ProposalSortDirection.ASC]: ProposalSortDirection.DESC,
+  [ProposalSortDirection.DESC]: ProposalSortDirection.ASC,
+};
+
+function sortProposalRowData(
+  sort: ProposalSortOptions,
+  data: ProposalRowData[],
+) {
+  if (sort.field === ProposalSortField.QUORUM) {
+    if (sort.direction === ProposalSortDirection.ASC) {
+      return data.slice().sort((a, b) => +b.currentQuorum - +a.currentQuorum);
+    } else {
+      return data.slice().sort((a, b) => +a.currentQuorum - +b.currentQuorum);
+    }
+  }
+  // safe to assume the desired sort field is voting ends column
+  // since there are only two sortable columns.
+  else {
+    if (sort.direction === ProposalSortDirection.ASC) {
+      return data.slice().sort((a, b) => {
+        const aTime = a.votingEnds ? a.votingEnds.getTime() : 0;
+        const bTime = b.votingEnds ? b.votingEnds.getTime() : 0;
+        return bTime - aTime;
+      });
+    } else {
+      return data.slice().sort((a, b) => {
+        const aTime = a.votingEnds ? a.votingEnds.getTime() : 0;
+        const bTime = b.votingEnds ? b.votingEnds.getTime() : 0;
+        return aTime - bTime;
+      });
+    }
   }
 }
